requirements.txt
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
pyyaml>=6.0
regex>=2023.0
streamlit>=1.35.0
